requirements.txt
django
djangorestframework
firebase_admin
requests
django_slack
fcm_django
celery
django-celery-beat
django-import-export
pytest
pytest-django
django-dbtemplates
django-render-block
